fix(db): descuentos FLOAT → NUMERIC(5,2) — evitar error de redondeo monetario (#645)
* fix(db): descuentos FLOAT → NUMERIC(5,2) para evitar error de redondeo monetario

Problema: los porcentajes de descuento se guardaban como FLOAT (double
precision). Un 10% almacenado como 0.09999999... aplicado a $5000 puede
dar $499.99 que redondea a $499 en vez de $500 — error de ±1 ARS por
operación, acumulable en pedidos con múltiples ítems.

Columnas migradas (4 de las 5 del plan; kit_componentes.descuento_pct
no existe en el schema actual de main):
- alquileres.descuento_pct
- alquileres.descuento_jornadas_pct
- clientes.descuento
- descuentos_jornada.pct

Cambios:
- Nueva migración Alembic g1a2b3c4d5e6 con upgrade() y downgrade() completos
- USING ROUND(col::NUMERIC, 2) para convertir valores existentes sin pérdida
- database.py: schema de init_db actualizado a NUMERIC(5,2) para nuevas instancias

psycopg2 convierte NUMERIC a Decimal() automáticamente — el código Python
que hace float(valor) sigue funcionando sin cambios (Decimal → float es
implícito en aritmética Python).


* fix(pr6-ci): corregir down_revision de la migración — apuntar al head real

El down_revision apuntaba a f9a3c5d8b1e7 (compatibilidad_modos) pero el
head real de la cadena era f1a9c7e3b5d2 (cliente_favoritos). Esto causaba
dos heads divergentes que rompían el test test_alembic_script_directory.


* fix(pr6-ci): down_revision → a1c3b5f7e9d2 (equipos_tipo, head real post-PR636)


---------

+++ backend/database.py
@@ -336,7 +336,7 @@ def init_db():
             email             TEXT NOT NULL UNIQUE,
             direccion         TEXT NOT NULL,
             cuit              TEXT NOT NULL,
-            descuento         FLOAT DEFAULT 0,
+            descuento         NUMERIC(5,2) DEFAULT 0,
             perfil_impuestos  TEXT DEFAULT 'consumidor_final',
             razon_social      TEXT,
             domicilio_fiscal  TEXT,
@@ -368,7 +368,7 @@ def init_db():
             fecha_hasta      TIMESTAMP,
             monto_total      INTEGER DEFAULT 0,
             monto_pagado     INTEGER DEFAULT 0,
-            descuento_pct    FLOAT DEFAULT 0,
+            descuento_pct    NUMERIC(5,2) DEFAULT 0,
             fuente           TEXT DEFAULT 'sistema',
             numero_pedido    INTEGER,
             created_at       TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
@@ -1065,7 +1065,7 @@ def init_db():
         CREATE TABLE IF NOT EXISTS descuentos_jornada (
             id         SERIAL PRIMARY KEY,
             jornadas   INTEGER NOT NULL UNIQUE,
-            pct        FLOAT   NOT NULL,
+            pct        NUMERIC(5,2) NOT NULL,
             created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
         )
     """)
